third_party/blink/web_tests/external/wpt/selection/contenteditable/collapse-expected.txt

This is a testharness.js-based test.
[FAIL] Selection.collapse() must succeed for the ancestor
  assert_equals: expected Element node <div contenteditable="" id="host3">\n  <div contenteditabl... but got Element node <div contenteditable="" id="host4"></div>
[FAIL] Selection.collapse() must succeed for the descendant
  assert_equals: expected Element node <div contenteditable="" id="host3">\n  <div contenteditabl... but got Element node <button id="button"></button>
Harness: the test ran to completion.
